+++ lib/nylas/send_grid_verified_status.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Nylas
+  # Ruby representation of a Nylas Send Grid verified status object
+  # @see https://docs.nylas.com/reference#drafts
+  class SendGridVerifiedStatus
+    include Model::Attributable
+
+    attribute :domain_verified, :boolean
+    attribute :sender_verified, :boolean
+  end
+end