+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Unreleased
+* Add Outbox support
 * Enable Nylas API v2.5 support
 
 ### 5.8.0 / 2022-03-18