+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 * Add Outbox support
 * Add new `authentication_type` field in Account
+* Add support for basic authentication
 * Enable Nylas API v2.5 support
 
 ### 5.8.0 / 2022-03-18