+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### Unreleased
+### 5.9.0 / 2022-04-07
 * Add Outbox support
 * Add new `authentication_type` field in Account
 * Add support for basic authentication